update fee type

+++ sol-parking-marketplace/programs/marketplace/src/state/marketplace.rs
@@ -1,13 +1,13 @@
 use anchor_lang::prelude::*;
 
 #[account]
-#[derive(InitSpace)]
+#[derive(InitSpace, Debug)]
 pub struct Marketplace {
     pub admin: Pubkey,
-    pub fee: u16, //basis points
+    pub fee: u32, //basis points
     pub bump: u8, 
    
-    #[max_len(32)] 
+    #[max_len(16)] //be careful of length when using name as seed, seeds total cannot exceed 32bytes
     pub name: String
 
 }
